BAU: add type hints
Type hints were introduced in Python 3.5 as a way of documenting the types used. They can be consumed by type checkers such as mypy (and other tools/linters/IDEs/etc.). Add type hints to our client code to make it slightly clearer what all the functions are doing and what their inputs are.

I have checked using mypy, which was happy with all the types. In a follow-up commit, I'll add this to the test suite.